fix: change ClientDocumentRepository finder year params from short to int
Avoids (short) casts at every DocumentService call site. Entity field
stays short for correct SMALLINT Hibernate mapping; Spring Data handles
the narrowing internally.

+++ backend/src/main/java/com/gwhaitech/accountingfirm/client/domain/ClientDocumentRepository.java
@@ -7,7 +7,7 @@ import java.util.Optional;
 
 public interface ClientDocumentRepository extends JpaRepository<ClientDocument, Long> {
 
-    List<ClientDocument> findByClientIdAndYear(Long clientId, short year);
+    List<ClientDocument> findByClientIdAndYear(Long clientId, int year);
 
-    Optional<ClientDocument> findByClientIdAndYearAndFilename(Long clientId, short year, String filename);
+    Optional<ClientDocument> findByClientIdAndYearAndFilename(Long clientId, int year, String filename);
 }
